requirements.txt
aider-install==0.1.2
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
ecdsa==0.19.0
Flask==3.0.0
Flask-Cors==4.0.0
google-api-core==2.24.0
google-auth==2.25.2
google-cloud-bigquery==3.14.1
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.6.1
MarkupSafe==3.0.2
openai==1.52.0
packaging==24.1
pillow==11.0.0
proto-plus==1.24.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0.2
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
stripe==7.10.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
uv==0.5.11
Werkzeug==3.1.3
anthropic==0.14.0
google-generativeai==0.3.2
wheel @ file:///opt/homebrew/Cellar/python%403.12/3.12.7_1/libexec/wheel-0.44.0-py3-none-any.whl#sha256=c01326144883c50263d6967d1145640f3bbc8dfab32c2cbe817ef61d7a810219
